# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Restless — owner cockpit API
  version: 0.1.0
  description: |
    The transport the operator SPA talks to. It is a **thin shim over the
    daemon's existing JSON-lines socket protocol**, not a separately designed
    API: every route builds the same request the `restless` CLI puts on the
    socket and runs it through the same `authorize` → `dispatch` path.

    That is deliberate. The CLI was already the API — a dumb client over a line
    protocol, with a principal on every request, typed errors, and one streaming
    command. Deriving a second surface from product concepts would give us two
    things to keep in step.

    ## The envelope

    Every response carries the same shape, success or failure:

    ```json
    { "ok": true,  "data": ... }
    { "ok": false, "error": { "kind": "authority", "message": "..." } }
    ```

    **Switch on `error.kind`, not on prose and not only on the status code.** An
    owner who sees `authority` knows to change what they are allowed to do; one
    who sees `transport` knows to check the daemon. HTTP has no code for that
    distinction, so the status is a second signal for proxies, never the only one.

    ## Authority

    The listener binds loopback only and stamps `principal: "owner"` — the same
    claim the unix socket makes, that a process on the host is the owner. It is a
    claim, not a proof. The stamp still travels through the daemon's `authorize`
...
    port a request arrived on.

...

...

...
servers:
...
paths:
...
  /v1/companies/{company}/status:
    get:
      tags: [read]
      summary: Company environment lifecycle status
      parameters: [{ $ref: '#/components/parameters/Company' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/people:
    get:
      tags: [read]
      summary: Who is in this company
      description: |
        Role, model, what each has cost, and whether a session is running right
        now. Cost is joined from the spend ledger rather than stored on the
        actor: OrgIntel owns role, the ledger owns cost, and neither becomes a
        second writer of the other.

        **Surface:** People (the directory).
      parameters: [{ $ref: '#/components/parameters/Company' }]
      responses:
        '200':
          description: One row per actor.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Person' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/goals:
    get:
      tags: [read]
      summary: Goals
      description: '**Surface:** Board (the roots of the tree).'
      parameters: [{ $ref: '#/components/parameters/Company' }]
      responses:
        '200':
          description: Goals, newest last.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Goal' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/commitments:
    get:
      tags: [read]
      summary: Commitments, all states
      description: |
        `goal_id` links a commitment to its goal, which is what makes the Board's
        tree real rather than drawn. A null `goal_id` is work that belongs to no
        goal — worth surfacing, not hiding.

        **Surface:** Board (tree children, and every kanban card).
      parameters: [{ $ref: '#/components/parameters/Company' }]
      responses:
        '200':
          description: Commitments.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Commitment' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/inbox:
    get:
      tags: [read]
      summary: Unread mail
      description: |
        **This read mutates.** It marks every message it returns as read, which
        breaks the safety a GET is supposed to have: a refresh or a prefetch
        silently destroys the owner's unread state. See `docs/api/MISSING.md` §6
        for the proposed split into a safe read plus `POST /inbox/seen`.

        Passing `as` inspects without marking — but it matches
        `to_actor = <actor>`, and the owner's own mail is stored with
        `to_actor IS NULL`, so `?as=owner` returns nothing rather than the
        owner's mail. There is currently no non-consuming read of the owner's
        inbox.

        **Surface:** Inbox (the message-shaped items), People (a person's mail).
      parameters:
        - { $ref: '#/components/parameters/Company' }
        - name: as
          in: query
          required: false
          schema: { type: string }
          description: Inspect this actor's inbox instead, without marking read.
      responses:
        '200':
          description: Messages.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Message' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/events:
    get:
      tags: [read]
      summary: Recent operational events, newest first
      description: '**Surface:** the executive dock, and any activity view.'
      parameters:
        - { $ref: '#/components/parameters/Company' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Events.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Event' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/receipts:
    get:
      tags: [read]
      summary: What the company actually did to the world
      description: |
        The strongest evidence the system holds about external effects: the
        capability, the provider, the party, the outcome, and the idempotency key
        that would replay it.

        **Surface:** none yet — nothing in the current design renders these, which
        is a gap in the design rather than in the API.
      parameters:
        - { $ref: '#/components/parameters/Company' }
        - name: capability
          in: query
          required: false
          schema: { type: string, examples: ['email.send'] }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Receipts, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Receipt' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/spend:
    get:
      tags: [read]
      summary: Spend against the ceiling
      description: |
        Note `poisoned`. A fail-closed accounting failure pins the ledger total at
        a sentinel so every preflight refuses; rendering that as money would print
        a fabricated figure. When `poisoned` is true, `remaining_usd` is **null**
        and `accounted_usd` is the real recorded spend. A UI must show the stopped
        state, not a number.

        **Surface:** People (a person's monthly spend), and any ceiling display.
      parameters: [{ $ref: '#/components/parameters/Company' }]
      responses:
        '200':
          description: Spend.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Spend' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/stream:
    get:
      tags: [read]
      summary: The operational event stream (SSE)
      description: |
        A snapshot of the last 20 events, then new ones as they land. Same poll,
        same watermark, and the same durable table as the CLI's `watch` — the
        stream survives the client's attention, not the other way around.

        Each `message` event's `data` is one `Event` object. A failure arrives as
        an SSE event named `error` whose data is the standard failure envelope.
      parameters: [{ $ref: '#/components/parameters/Company' }]
      responses:
        '200':
          description: An endless SSE stream.
          content:
            text/event-stream:
              schema: { type: string }
  /v1/companies/{company}/up:
    post:
      tags: [write]
      summary: Start the company's computer
      description: |
        Creates the volume and container if absent, starts it if stopped, seeds
        the mission onto the volume, and ensures the OrgIntel schema. Separate
        from creation on purpose: a launch that fails must not lose the company
        the owner just described.

        `reconcile: true` rebuilds the image first and takes **minutes** — it
        compiles Rust and installs a browser. Prefer `restless up --reconcile`
        from the CLI for that; this route defaults it to false.

        Owner-only.

        **Surface:** onboarding.
      parameters: [{ $ref: '#/components/parameters/Company' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reconcile: { type: boolean, default: false }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/Refused' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/tell:
    post:
      tags: [write]
      summary: Issue a directive to the Exec
      description: |
        Wakes the company. Also how a blocked judgement request is answered.

        **Surface:** the executive dock composer.
      parameters: [{ $ref: '#/components/parameters/Company' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, description: What to tell the Exec. }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Failed' }
  /v1/companies/{company}/wake:
    post:
      tags: [write]
      summary: Wake the Exec for one turn
      parameters: [{ $ref: '#/components/parameters/Company' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: [string, 'null']
                  description: Why this wake is happening. Absent means owner-requested.
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Failed' }
...
  /v1/companies/{company}/attention:
    get:
      tags: [stub]
      summary: 'STUB — the merged attention stack'
      description: |
        **Not implemented. Returns `data: null`.**

        **Surface:** Inbox. See `docs/api/MISSING.md`.
      parameters: [{ $ref: '#/components/parameters/Company' }]
      responses:
        '200': { $ref: '#/components/responses/Stub' }
...
components:
  parameters:
    Company:
      name: company
      in: path
      required: true
      schema: { type: string }
      description: Company name — also its container, volume and OrgIntel schema.
...
  responses:
    Ok:
      description: Succeeded.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...